+++ SpatialGDK/Source/SpatialGDK/Public/Schema/NetOwningClientWorker.h
@@ -19,8 +19,8 @@ struct NetOwningClientWorker : Component
 
 	NetOwningClientWorker() = default;
 
-	NetOwningClientWorker(const TSchemaOption<PhysicalWorkerName>& InWorkerId)
-		: WorkerId(InWorkerId)
+	NetOwningClientWorker(const TSchemaOption<Worker_PartitionId>& InPartitionId)
+		: ClientPartitionId(InPartitionId)
 	{
 	}
 
@@ -32,45 +32,45 @@ struct NetOwningClientWorker : Component
 	NetOwningClientWorker(Schema_ComponentData* Data)
 	{
 		Schema_Object* ComponentObject = Schema_GetComponentDataFields(Data);
-		if (Schema_GetBytesCount(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID) == 1)
+		if (Schema_GetEntityIdCount(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID) == 1)
 		{
-			WorkerId = GetStringFromSchema(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID);
+			ClientPartitionId = Schema_GetEntityId(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID);
 		}
 	}
 
-	Worker_ComponentData CreateNetOwningClientWorkerData() { return CreateNetOwningClientWorkerData(WorkerId); }
+	Worker_ComponentData CreateNetOwningClientWorkerData() { return CreateNetOwningClientWorkerData(ClientPartitionId); }
 
-	static Worker_ComponentData CreateNetOwningClientWorkerData(const TSchemaOption<PhysicalWorkerName>& WorkerId)
+	static Worker_ComponentData CreateNetOwningClientWorkerData(const TSchemaOption<Worker_PartitionId>& PartitionId)
 	{
 		Worker_ComponentData Data = {};
 		Data.component_id = ComponentId;
 		Data.schema_type = Schema_CreateComponentData();
 		Schema_Object* ComponentObject = Schema_GetComponentDataFields(Data.schema_type);
 
-		if (WorkerId.IsSet())
+		if (PartitionId.IsSet())
 		{
-			AddStringToSchema(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID, *WorkerId);
+			Schema_AddEntityId(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID, *PartitionId);
 		}
 
 		return Data;
 	}
 
-	Worker_ComponentUpdate CreateNetOwningClientWorkerUpdate() { return CreateNetOwningClientWorkerUpdate(WorkerId); }
+	Worker_ComponentUpdate CreateNetOwningClientWorkerUpdate() const { return CreateNetOwningClientWorkerUpdate(ClientPartitionId); }
 
-	static Worker_ComponentUpdate CreateNetOwningClientWorkerUpdate(const TSchemaOption<PhysicalWorkerName>& WorkerId)
+	static Worker_ComponentUpdate CreateNetOwningClientWorkerUpdate(const TSchemaOption<Worker_PartitionId>& PartitionId)
 	{
 		Worker_ComponentUpdate Update = {};
 		Update.component_id = ComponentId;
 		Update.schema_type = Schema_CreateComponentUpdate();
 		Schema_Object* ComponentObject = Schema_GetComponentUpdateFields(Update.schema_type);
 
-		if (WorkerId.IsSet())
+		if (PartitionId.IsSet())
 		{
-			AddStringToSchema(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID, *WorkerId);
+			Schema_AddEntityId(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID, *PartitionId);
 		}
 		else
 		{
-			Schema_AddComponentUpdateClearedField(Update.schema_type, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID);
+			Schema_AddComponentUpdateClearedField(Update.schema_type, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID);
 		}
 
 		return Update;
@@ -81,18 +81,24 @@ struct NetOwningClientWorker : Component
 	void ApplyComponentUpdate(Schema_ComponentUpdate* Update)
 	{
 		Schema_Object* ComponentObject = Schema_GetComponentUpdateFields(Update);
-		if (Schema_GetBytesCount(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID) == 1)
+
+		if (Schema_GetEntityIdCount(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID) == 1)
 		{
-			WorkerId = GetStringFromSchema(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID);
+			ClientPartitionId = Schema_GetEntityId(ComponentObject, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID);
 		}
-		else if (Schema_IsComponentUpdateFieldCleared(Update, SpatialConstants::NET_OWNING_CLIENT_WORKER_FIELD_ID))
+		else if (Schema_IsComponentUpdateFieldCleared(Update, SpatialConstants::NET_OWNING_CLIENT_PARTITION_ENTITY_FIELD_ID))
 		{
-			WorkerId = TSchemaOption<PhysicalWorkerName>();
+			ClientPartitionId = TSchemaOption<Worker_PartitionId>();
 		}
 	}
 
-	// Client worker ID corresponding to the owning net connection (if exists).
-	TSchemaOption<PhysicalWorkerName> WorkerId;
+	void SetPartitionId(const Worker_PartitionId& InPartitionId)
+	{
+		ClientPartitionId = InPartitionId == SpatialConstants::INVALID_ENTITY_ID ? TSchemaOption<Worker_PartitionId>() : InPartitionId;
+	}
+
+	// Client partition entity ID corresponding to the owning net connection (if exists).
+	TSchemaOption<Worker_PartitionId> ClientPartitionId;
 };
 
 } // namespace SpatialGDK